# Optimize `list.c`