--- a/lib/module.stk
+++ b/lib/module.stk
@@ -431,7 +431,7 @@ doc>
     (let* ((mod       (eventually-load-library name))
            (imported  (build-import-list mod
                                          keylist
-                                         (list-copy (%module-exports mod)))))
+                                         (list-deep-copy (%module-exports mod)))))
       ;; Make aliases for all symbols in imported list
       (if only-syntax
           (for-each (lambda (pair)
@@ -476,10 +476,10 @@ doc>
                 (%grab-file-information (symbol->string (car x))))
               imp)
     ;; Do importation at compile time AND at execution time
-    (%do-imports (compiler-current-module) (list-copy imp) #t)
+    (%do-imports (compiler-current-module) (list-deep-copy imp) #t)
 
     ;; Run time expansion
-    `(%do-imports (current-module) (list-copy ',imp) #f)))
+    `(%do-imports (current-module) (list-deep-copy ',imp) #f)))
 
 (define (%%import lib modules)
   (let ((imp (%parse-imports modules)))
@@ -490,7 +490,7 @@ doc>
               imp)
     ;; Do importation at compile time AND at execution time
     ;(%do-imports  (compiler-current-module) (list-copy imp) #t)
-    (%do-imports lib (list-copy imp) #f)))
+    (%do-imports lib (list-deep-copy imp) #f)))
 
 
 ;;=============================================================================