--- a/lib/scheme/list.stk
+++ b/lib/scheme/list.stk
@@ -694,11 +694,12 @@
 
 (define (last lis) (car (last-pair lis)))
 
-(define (last-pair lis)
-  (check-arg pair? lis last-pair)
-  (let lp ((lis lis))
-    (let ((tail (cdr lis)))
-      (if (pair? tail) (lp tail) lis))))
+;; Already in STklos core
+;; (define (last-pair lis)
+;;   (check-arg pair? lis last-pair)
+;;   (let lp ((lis lis))
+;;     (let ((tail (cdr lis)))
+;;       (if (pair? tail) (lp tail) lis))))
 
 
 ;;; Unzippers -- 1 through 5
@@ -1772,6 +1773,7 @@ FILTER an FILTER! are primitives in STklos
 (redefine member)
 
 ;; Not R7RS, but also moved to STklos core:
+(redefine last-pair)
 (redefine circular-list?)
 
 )  ; END OF MODULE