--- a/src/str.c
+++ b/src/str.c
@@ -939,22 +939,19 @@ DEFINE_PRIMITIVE("string->list", string2list, subr1, (SCM str))
   register char *s;
   int len;
   utf8_char c;
-  SCM tmp, tmp1, z;
+  SCM tmp, z;
 
   if (!STRINGP(str)) error_bad_string(str);
 
   len = STRING_LENGTH(str);
   s   = STRING_CHARS(str);
 
-  tmp = z = STk_nil;
+  tmp = z = STk_must_malloc_list(len,STk_false);
 
-  while (len--) {
+  for (int i=0; i<len; i++) {
     s = STk_utf8_grab_char(s, &c);
-    tmp1 = STk_cons(MAKE_CHARACTER(c), STk_nil);
-    if (z == STk_nil)
-      tmp = z = tmp1;
-    else
-      tmp = CDR(tmp) = tmp1;
+    CAR(tmp) = MAKE_CHARACTER(c);
+    tmp = CDR(tmp);
   }
   return z;
 }