… src/vector.c