Hi @egallesio !
These are several enhancements to `list.c` and related functions in `proc.c`.

Passes all tests. :)

And there are some really big performance gains, besides some enhancements in argument checking and error messages. (This brings the new allocator for lists, which does not repeatedly cons, but asks for a large set of cells each time)

# Optimize make-list and vector->list

Allocate more cells at once in `make-list`.

`libgc` only offers an interface to "get several objects", without telling us how many. It's usually a not very large number like 128 objects for example). We give the size, and receive a list.

This patch does this:

1. Implements `SCM STk_must_malloc_list(int n, SCM init)` that will allocate a list of CONS cells.
2. Changes `vector->list` so it will use the new function
3. Removes `make-list` and use the new function instead

There is also an internal function `STk_must_malloc_list_upton`, which allocates a list of size "at most n". This function will receive a size argument `n`, and will call the `libgc` function to allocate many objects.
1. If there is more than `n`, it will run through the list, cleaning the pointers so they are released.
2. If there is less than `n`, it will return the number of missing cells (not allocated) in the parameter `n`

`STk_must_malloc_list`  does:
1. calls `STk_must_malloc_list_upton`
3. If there are enough cells, stop
4. If not, call  `STk_must_malloc_list_upton` again, and append the two lists. Appending is NOT linear, it is O(1) here, because `STk_must_malloc_list_upton` returns a pointer to the last cell :)

New timings:
```scheme
(time
  (let ((a (make-vector 1000 'x)))
    (repeat 100000 (vector->list a))
    'ok))
```
Old code: 4152.755ms 100000001 allocation calls
New code: 2552.173ms  934202 allocation calls

```scheme
(time
  (let ((a #f))
    (repeat 100000
      (set! a (make-list 1000)))
    'ok))
```
Old code: 5339.514 ms 100100000 allocation calls
New code:  2249.643 ms 935418 allocation calls

Now make 10x less lists, with 10x more elements per list:
```scheme
(time
  (let ((a #f))
    (repeat 10000
      (set! a (make-list 10000)))
    'ok))
```
Old code: 4928.766 ms 100010000 allocation calls
New code: 2706.778 ms 873160 allocation calls

Also -- one more thing.

The new code properly counts allocation.

The size of `struct cons_object` on my machine is `24`.
```c

typedef struct {
  int16_t type, cell_info;
} stk_header;

struct cons_obj {
  stk_header header;
  void * car;
  void* cdr;
};
int main() {
	printf("size is %ld\n", sizeof(struct cons_obj));
}
```
The above program prints 24.

So -- with the new code,

```scheme
(time (make-list 10))
;    Elapsed time: 0.017 ms
;    Allocations: 240 bytes in 1 allocation call (GC: 0)

(time (make-list 10000))
;    Elapsed time: 0.131 ms
;    Allocations: 240000 bytes in 116 allocation calls (GC: 0)
```
**Exactly** the requested number of bytes (ok, we're cheating, because we allocated more and threw hem away, but they're gone immediately after allocation).

The old code, for some reason, was allocating more than necessary:
```scheme
(time (make-list 10))
;    Elapsed time: 0.007 ms
;    Allocations: 328 bytes in 11 allocation calls (GC: 0)

(time (make-list 10000))
;    Elapsed time: 0.955 ms
;    Allocations: 320008 bytes in 10001 allocation calls (GC: 0)
```
# Optimize MAP

Allocate the result list at once instead of consing all of the result list.

```scheme
(import (scheme list))

(let ((L (iota 30000000)))
  (time (begin (map - L)
               -1)))
```
Old code: 1728.254 ms
New code: 1131.081 ms

```scheme
(let ((L (iota 20000000))
      (M (reverse (map - (iota 18000000)))))
  (time (begin (map * L M)
               -1)))
```
Old code: 1273.321 ms
New code: 1021.081 ms

# Optimize and enhance list-copy, add list-deep-copy

1. Use Floyd's algorithm to detect cycles. It now copies:
   - proper lists
   - improper lists
   - cyclic lists
2. Allocate the list at once, not cell-by-cell
3. Do not use recursion. This avoids crashes with large lists
4. For all this to work, we implement a new C function,  `STk_list_type_and_length()`, which
   - tells wether the object is a proper list, improper list,  cyclic list or no list at all;
   - calculates the length of the list (if it is one)
5. Since we did all that, move circular-list? from SRFI-1  to the core
6. Add list-deep-copy. This is necessary in the module import  code. The previous version of `list-copy` would recurse    into the list, and the new won't. Reasons for making it    different, and create the separate `list-deep-copy`:
   - Every Scheme except STklos behaves the same way (they  won't recurse into the list); Gauche even has a  specialized `alist-copy`.
   - The `list-copy` primitiveis much faster than the old  one

See comments in the code.

```scheme
(let ((L (iota 20000)))
  (time (begin (repeat 3000 (list-copy L))
               -1)))
```
Old code: 2157.288 ms
New code: 1590.699 ms

```scheme
(let ((L (iota 20000000)))
  (begin (list-copy L)
          -1))
```
Old code crashed
New code works fine (The C function doesn't do recursive calls).

# Optimize and enhance reverse

`reverse` is a bit faster now, since it calls list-copy to allocate cells all at once.

And now STklos tells when a list is circular or improper:

stklos> (reverse '(1 . 2))
**** Error:
reverse: improper list '(1 . 2)'

stklos> (reverse (circular-list 1 2 3))
**** Error:
reverse: list '#0=(1 2 3 . #0#)' is circular

stklos> (reverse (append '(11 22) (circular-list 1 2 3)))
**** Error:
reverse: list '(11 22 . #0=(1 2 3 . #0#))' is circular
	(type ",help" for more information)

The two last examples would make STklos enter an infinite loop before this change.

#  Optimize and enhance `last-pair`

There was already a `last-pair` primitive (?) in STklos, but there was also a closure overriding it.
We enhanced the primitive a bit, and made it be used by default.

# Enhance `filter`

It now complains more clearly about circular and improper lists.