@jpellegrini 2025-03-16
By the way... I know that R7RS says "it is an error" if the list passed to `list-copy` is circular. But it doesn't require us to signal an error, and we can do whatever we think is best... :smile: