@jpellegrini 2025-03-17
I'll probably do more enhancements to the way libgc is used... In the near future. But this is already some improvement, no? :)